fix: stop naming the engines uf drives internally
The product decision is that `uf` borrows the plugin and task semantics of the
existing ecosystem so those plugins keep working, and does not borrow the names.
A developer using `uf` writes `uf.config.js` and `.js` files with `// @flow`;
they never chose a bundler, and nothing should hand them one to reason about.

Eight places still did, and they were the ones that matter most — not comments,
but strings a user reads:

- `dist/uf-build-manifest.json` shipped `"bundlerCompatibility": ["vite", "rolldown"]`
  into their repository, and `.uf/dev-server.json` shipped two more. Both now
  carry `"pluginContract": "uf-plugin-v1"`, which is the thing a reader actually
  needs: which contract the manifest speaks.
- `uf inspect` reported the build and dev server as `vite-compatible/rolldown`.
- A failed task said it ran "through uf Vite Task runner".
- `taskRunner.engine` was spelled `"vite-task"` in `uf.config.js` and in the Flow
  type for it. It is `"uf-task"` now, with **no alias** for the old spelling: a
  name a user can still write is still a name they can see, which is the thing
  being removed.

`uf_plugin`'s naming guard already read its own sources and the project
templates. Widen it to `uf_cli` (it writes both the terminal output and the
manifests that land in `dist/`), `uf_config` (its enum spellings *are* the
vocabulary of `uf.config.js`), and the shipped `@uniflowed/*` modules, whose
comments a user reads in their own `node_modules` — one of which described
`defineConfig` by comparison to the entry point it replaces.

`#[cfg(test)]` code stays exempt, now as a `tests.rs` sibling as well as a
`tests` directory. A test for "this name never reaches a user" has to say the
name, and a config fixture may legitimately hold a *user's* task command such as
`vite --host 0.0.0.0` — that string is theirs, not ours.

`uf_config`'s inline test module moves to `src/tests.rs` so the guard can cover
the crate, which also brings `lib.rs` from 2178 lines to 1779.

`TaskRunnerEngine` becomes `#[non_exhaustive]`, so a second runner is additive.
